docs/hosts: remove hint, discuss replit / heroku
Generally speaking, it is technically possible to host YAGPDB on repl.it
and/or Heroku, but it is strongly discouraged.

Their free plans do not grant the necessary data integrity and access
which are implicitly required to pull up a YAGPDB instance and maintain
it.
This commit adds a brief discussion of the ins and outs of using such
free services, and specifically mentions the issue regarding those free
plans.